"Save" for fan gui, device database update, small fixes

+++ alienfx-gui/FanDialog.cpp
@@ -222,7 +222,7 @@ BOOL CALLBACK TabFanDialog(HWND hDlg, UINT message, WPARAM wParam, LPARAM lParam
                 NMLVDISPINFO* sItem = (NMLVDISPINFO*)lParam;
                 fanUIUpdate->Stop();
                 HWND editC = ListView_GetEditControl(tempList);
-                auto pwr = fan_conf->sensors.find(sItem->item.lParam);
+                auto pwr = fan_conf->sensors.find((WORD)sItem->item.lParam);
                 Edit_SetText(ListView_GetEditControl(tempList), (pwr != fan_conf->sensors.end() ? pwr->second : acpi->sensors[sItem->item.iItem].name).c_str());
             } break;
             case LVN_ITEMACTIVATE: case NM_RETURN:
@@ -236,10 +236,10 @@ BOOL CALLBACK TabFanDialog(HWND hDlg, UINT message, WPARAM wParam, LPARAM lParam
             {
                 NMLVDISPINFO* sItem = (NMLVDISPINFO*)lParam;
                 if (sItem->item.pszText) {
-                    auto pwr = fan_conf->sensors.find(sItem->item.lParam);
+                    auto pwr = fan_conf->sensors.find((WORD)sItem->item.lParam);
                     if (pwr == fan_conf->sensors.end()) {
                         if (strlen(sItem->item.pszText))
-                            fan_conf->sensors.emplace(sItem->item.lParam, sItem->item.pszText);
+                            fan_conf->sensors.emplace((WORD)sItem->item.lParam, sItem->item.pszText);
                     }
                     else {
                         if (strlen(sItem->item.pszText))
